requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.1
aiohttp==3.9.1
qrcode==7.4.2
pillow==10.1.0
python-multipart==0.0.6
pydantic[email]
requests==2.31.0
langchain==0.2.17
openai==1.109.1 
PyPDF2==3.0.1
python-docx==1.2.0
langchain-openai==0.1.25

pandas